Separator files out into responsibilities
