+++ generate_wiki/readme.py
@@ -1,0 +1,6 @@
+from generate_wiki.repo import Repo
+
+
+class ReadMeCatalog:
+    def __init__(self, repo):
+        self.repo: Repo = repo
